requirements.txt
# PyTorch + CUDA 12.1
torch==2.3.0+cu121
torchvision==0.18.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Przetwarzanie obrazów i dane
numpy<2.0
pillow
matplotlib

# Do zapisu modeli i śledzenia wyników
tqdm
pillow
torchcam
torchviz
streamlit

# (Opcjonalnie) Notatniki Jupyter
notebook
